+++ crates/turbopack-cli-utils/Cargo.toml
@@ -15,6 +15,7 @@ bench = false
 anyhow = { workspace = true }
 clap = { workspace = true, features = ["derive"] }
 crossterm = "0.26.0"
+ctrlc = "3.2.5"
 owo-colors = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
