+++ crates/turbopack-cli/Cargo.toml
@@ -40,7 +40,6 @@ anyhow = { workspace = true, features = ["backtrace"] }
 clap = { workspace = true, features = ["derive", "env"] }
 console-subscriber = { workspace = true, optional = true }
 criterion = { workspace = true, features = ["async_tokio"] }
-ctrlc = "3.2.5"
 dunce = { workspace = true }
 futures = { workspace = true }
 mime = { workspace = true }
