+++ crates/turbopack-cli/src/main.rs
@@ -1,45 +1,21 @@
 #![feature(future_join)]
 #![feature(min_specialization)]
 
-use std::{
-    borrow::Cow,
-    sync::{Arc, Mutex},
-};
+use std::{borrow::Cow, path::Path};
 
 use anyhow::{Context, Result};
 use clap::Parser;
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt, EnvFilter, Registry};
 use turbopack_cli::{arguments::Arguments, register};
-use turbopack_cli_utils::raw_trace::RawTraceLayer;
+use turbopack_cli_utils::{exit::exit_guard, raw_trace::RawTraceLayer};
 
 #[global_allocator]
 static ALLOC: turbo_tasks_malloc::TurboMalloc = turbo_tasks_malloc::TurboMalloc;
 
-struct CloseGuard<T>(Arc<Mutex<Option<T>>>);
-
-impl<T> Drop for CloseGuard<T> {
-    fn drop(&mut self) {
-        drop(self.0.lock().unwrap().take())
-    }
-}
-
-fn close_guard<T: Send + 'static>(guard: T) -> Result<CloseGuard<T>> {
-    let guard = Arc::new(Mutex::new(Some(guard)));
-    {
-        let guard = guard.clone();
-        ctrlc::set_handler(move || {
-            println!("Flushing trace file... (ctrl-c)");
-            drop(guard.lock().unwrap().take());
-            println!("Flushed trace file");
-            std::process::exit(0);
-        })
-        .context("Unable to set ctrl-c handler")?;
-    }
-    Ok(CloseGuard(guard))
-}
-
 fn main() {
     use turbo_tasks_malloc::TurboMalloc;
+
+    let args = Arguments::parse();
 
     let subscriber = Registry::default();
 
@@ -62,14 +38,19 @@ fn main() {
             .unwrap(),
     );
 
-    std::fs::create_dir_all("./.turbopack")
+    let internal_dir = args
+        .dir()
+        .unwrap_or_else(|| Path::new("."))
+        .join(".turbopack");
+    std::fs::create_dir_all(&internal_dir)
         .context("Unable to create .turbopack directory")
         .unwrap();
+    let trace_file = internal_dir.join("trace.log");
     let (writer, guard) =
-        tracing_appender::non_blocking(std::fs::File::create("./.turbopack/trace.log").unwrap());
+        tracing_appender::non_blocking(std::fs::File::create(trace_file).unwrap());
     let subscriber = subscriber.with(RawTraceLayer::new(writer));
 
-    let guard = close_guard(guard).unwrap();
+    let guard = exit_guard(guard).unwrap();
 
     subscriber.init();
 
@@ -80,17 +61,14 @@ fn main() {
         })
         .build()
         .unwrap()
-        .block_on(main_inner())
+        .block_on(main_inner(args))
         .unwrap();
 
-    println!("Flushing trace file...");
     drop(guard);
-    println!("Flushed trace file");
 }
 
-async fn main_inner() -> Result<()> {
+async fn main_inner(args: Arguments) -> Result<()> {
     register();
-    let args = Arguments::parse();
 
     match args {
         Arguments::Dev(args) => turbopack_cli::dev::start_server(&args).await,
